#20: Create a subcomponent to re-use it
